ref(ui): Move `getCookie` to module
